HOSTEDCP-336: Move fixtures out of API
