+++ cmd/nodepool/kubevirt/create.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/openshift/hypershift/api/fixtures"
+	"github.com/openshift/hypershift/examples/fixtures"
 	"github.com/spf13/cobra"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
