+++ examples/fixtures/example_kubevirt.go
